Dirty hack to keep me muted
This is because for some reason the listeners are not firing when Zoom unmutes my mic.
I dont think this is my bug as even the OSX system settings UI seems to have the volume at 0.0 when this happens.

OSX bug??
Is there anything else I can listen to instead and remain event driven??